+++ config/services.php
@@ -43,4 +43,10 @@ return [
         'key' => env('ANTHROPIC_API_KEY'),
     ],
 
+    'twilio' => [
+        'sid' => env('TWILIO_SID'),
+        'token' => env('TWILIO_TOKEN'),
+        'from' => env('TWILIO_FROM'),
+    ],
+
 ];